Add API for setting HDF5 cache size
