+++ MDANSE/Src/MDANSE/MolecularDynamics/Trajectory.py
@@ -262,6 +262,11 @@ class Trajectory:
         dataset_type = TrajDataArray[array_name.upper()]
         return self._trajectory.chunk_size(dataset_type)
 
+    def dtype_size(self, array_name: str = "position") -> int:
+        """Return the number of atoms in a single chunk of the HDF5 dataset."""
+        dataset_type = TrajDataArray[array_name.upper()]
+        return self._trajectory.dtype_size(dataset_type)
+
     def group_elements(self, grp_name):
         """The elements in the group with the name grp_name.
 
@@ -501,7 +506,7 @@ class Trajectory:
 
     def __setstate__(self, state):
         self.__dict__ = state
-        self._trajectory = self.open_trajectory(self._format, self._hdf5_driver, self._dataset_cache_size)
+        self._trajectory = self.open_trajectory(self._format, self._hdf5_driver, self._rdcc_nbytes, self._rdcc_w0, self._rdcc_nslots)
 
     def __len__(self):
         return len(self._trajectory)
