+++ MDANSE/Src/MDANSE/Trajectory/FileTrajBase.py
@@ -24,6 +24,7 @@ import numpy as np
 import numpy.typing as npt
 
 from MDANSE.Chemistry import ATOMS_DATABASE
+from MDANSE.MLogging import LOG
 from MDANSE.MolecularDynamics.TrajectoryUtils import (
     atomic_trajectory,
     atomic_trajectory_many,
@@ -131,10 +132,35 @@ class TrajectoryFile(ABC):
     @abstractmethod
     def time(self) -> FloatArray: ...
 
-    @abstractmethod
-    def chunk_size(
-        self, dataset_type: TrajDataArray = TrajDataArray.POSITION
-    ) -> int: ...
+    def chunk_size(self, dataset_type: TrajDataArray = TrajDataArray.POSITION) -> int:
+        data_key = self.KEYS[dataset_type.name.lower()]
+        try:
+            dataset = self._h5_file[data_key]
+        except KeyError:
+            LOG.error("Dataset %s was not in the trajectory file", data_key)
+            return -1
+        if (chunk_shape := getattr(dataset, "chunks", None)) is None:
+            LOG.warning("Dataset %s is not chunked, and was expected to be", data_key)
+            return -1
+        if len(chunk_shape) < 2:
+            LOG.warning("Dataset %s does not have enough dimensions", data_key)
+            return -1
+        return chunk_shape[1]
+
+    def dtype_size(self, dataset_type: TrajDataArray = TrajDataArray.POSITION) -> int:
+        data_key = self.KEYS[dataset_type.name.lower()]
+        dataset = self._h5_file[data_key]
+        match dataset.dtype:
+            case np.float16:
+                return 2
+            case np.float32:
+                return 4
+            case np.float64:
+                return 8
+            case np.float128:
+                return 16
+            case _:
+                return 8
 
     @abstractmethod
     def unit_cell(self, frame: int) -> UnitCell | None: ...
